# docs/swagger.yaml
basePath: /
definitions:
...
  entity.Topic:
    properties:
      author_id:
        type: integer
      category_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  response.IDResponse:
    properties:
      id:
        example: 123
        type: integer
    type: object
...
  response.SuccessMessageResponse:
    properties:
      message:
        example: operation was successful
        type: string
    type: object
  response.TopicResponse:
    properties:
      topic:
        $ref: '#/definitions/entity.Topic'
    type: object
  response.TopicsResponse:
    properties:
      topics:
        items:
          $ref: '#/definitions/entity.Topic'
        type: array
    type: object
  topicrequests.UpdateRequest:
    properties:
      title:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: API for forum service
  title: Forum Service API
  version: "1.0"
paths:
...
  /categories/{id}/topics:
    get:
      description: Retrieves a list of topics for a category ID.
      parameters:
      - description: Category ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved topics
          schema:
            $ref: '#/definitions/response.TopicsResponse'
        "400":
          description: Invalid category ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get topics by category ID
      tags:
      - topics
    post:
      consumes:
      - application/json
      description: Creates a new topic in a category. Requires authentication.
      parameters:
      - description: Category ID to create topic in
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: Topic data to create. ID, AuthorID, CategoryID, CreatedAt, UpdatedAt
          will be ignored or overridden.
        in: body
        name: topic
        required: true
        schema:
          $ref: '#/definitions/entity.Topic'
      produces:
      - application/json
      responses:
        "200":
          description: Topic created successfully
          schema:
            $ref: '#/definitions/response.IDResponse'
        "400":
          description: Invalid category ID or request payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized (token is missing or invalid)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden (user is not authorized or trying to impersonate)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new topic
      tags:
      - topics
...
  /topics/{id}:
    delete:
      description: Deletes a topic by its ID. Requires authentication and ownership
        or role.
      parameters:
      - description: Topic ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Topic deleted successfully
          schema:
            $ref: '#/definitions/response.SuccessMessageResponse'
        "400":
          description: Invalid topic ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized (token is missing or invalid)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden (user is not an owner or admin)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Topic not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a topic
      tags:
      - topics
    get:
      description: Retrieves a specific topic by its ID.
      parameters:
      - description: Topic ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved topic
          schema:
            $ref: '#/definitions/response.TopicResponse'
        "400":
          description: Invalid topic ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to get topic
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get a topic by ID
      tags:
      - topics
    patch:
      consumes:
      - application/json
      description: Updates a topic. Requires authentication and ownership or admin
        role.
      parameters:
      - description: Topic ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: Topic update data (only title)
        in: body
        name: topic_update
        required: true
        schema:
          $ref: '#/definitions/topicrequests.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Topic updated successfully
          schema:
            $ref: '#/definitions/response.SuccessMessageResponse'
        "400":
          description: Invalid topic ID or request payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized (token is missing or invalid)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden (user is not an owner or admin)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Topic not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a topic
      tags:
      - topics
...
swagger: "2.0"